Function inquire::prompt_u64

pub fn prompt_u64<M>(message: M) -> InquireResult<u64>
where
    M: AsRef<str>,

This function is a helpful one-liner to prompt the user for a number and parse it to u64.

Under the hood, it is equivalent to calling inquire::CustomType::<u64>::new(message).prompt(). See the documentation for [inquire::CustomType] for more information on its behavior.

Arguments

message: A message that implements the AsRef<str> trait. This message will be displayed to the user when asking for input.

Returns

InquireResult<u64>: An enum that represents the result of the prompt operation. If the operation is successful, it returns InquireResult::Ok(u64) where u64 is the number parsed from the user’s input. If the operation encounters an error, it returns InquireResult::Err(InquireError).

Example

let kilograms = prompt_u64("Weight (kg):")?;

Errors

This function will return an error if there is a problem interacting with the terminal, or if the user cancels the operation by pressing Ctrl+C.