inquire

Trait InnerAction

pub trait InnerAction
where
    Self: Sized + Copy + Clone + PartialEq + Eq,
{
    type Config;

    // Required method
    fn from_key(key: Key, config: &Self::Config) -> Option<Self>
       where Self: Sized;
}

InnerActions are specialized prompt actions.

They must provide an implementation to optionally derive an action from a key event.

Required Associated Types

type Config

Configuration type for the prompt.

This is used to derive the action from a key event.

Required Methods

fn from_key(key: Key, config: &Self::Config) -> Option<Self>
where Self: Sized,

Derives a prompt action from a Key event and the prompt configuration.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.