Module inquire::ui

UI-related definitions for rendered content.

Structs

Attributes

Attributes to apply to a text via the StyleSheet struct.

ErrorMessageRenderConfig

Render configuration for error messages.

RenderConfig

Rendering configuration that can be applied to a prompt.

StyleSheet

Style definitions that can be applied to the rendered content.

Styled

Represents a content that when rendered must have the associated style applied to it.

Enums

Color

Represents a color to be used for text styling purposes.

IndexPrefix

Definition of index prefixes in option lists.