inquire::ui

Struct Styled

pub struct Styled<T>
where
    T: Display,
{
    pub content: T,
    pub style: StyleSheet,
}

Represents a content that when rendered must have the associated style applied to it.

Fields
content: T

Content to be rendered.

style: StyleSheet

Style sheet to be applied to content when rendered.

Implementations

impl<T> Styled<T>
where T: Display,

pub fn new(content: T) -> Self

Creates a new Styled object with the specified content and a default (empty) style sheet.

pub fn with_style_sheet(self, style_sheet: StyleSheet) -> Self

Sets the style sheet to the styled struct.

pub fn with_fg(self, fg: Color) -> Self

Sets the styled content to have the defined foreground Color.

pub fn with_bg(self, bg: Color) -> Self

Sets the styled content to have the defined foreground Color.

pub fn with_attr(self, attributes: Attributes) -> Self

Sets the styled content to have the defined attributes.

Warning: this does not keep the previously applied attributes. If you want to just set a new attribute and keep the others, you need to apply the OR operation yourself.

pub fn with_content<U>(self, content: U) -> Styled<U>
where U: Display,

Updates the content while keeping the style sheet constant.