inquire

Struct Password

pub struct Password<'a> {
    pub message: &'a str,
    pub custom_confirmation_message: Option<&'a str>,
    pub custom_confirmation_error_message: Option<&'a str>,
    pub help_message: Option<&'a str>,
    pub formatter: StringFormatter<'a>,
    pub display_mode: PasswordDisplayMode,
    pub enable_display_toggle: bool,
    pub enable_confirmation: bool,
    pub validators: Vec<Box<dyn StringValidator>>,
    pub render_config: RenderConfig<'a>,
}

Prompt meant for secretive text inputs.

By default, the password prompt behaves like a standard one youโd see in common CLI applications: the user has no UI indicators about the state of the current input. They do not know how many characters they typed, or which character they typed, with no option to display the current text input.

However, you can still customize these and other behaviors if you wish:

Standard display mode: Set the display mode of the text input among hidden, masked and full via the PasswordDisplayMode enum.

Hidden: default behavior, no UI indicators.

Masked: behaves like a normal text input, except that all characters of the input are masked to a special character, which is '*' by default but can be customized via RenderConfig.

Full: behaves like a normal text input, no modifications.

Toggle display mode: When enabling this feature by calling the with_display_toggle_enabled() method, you allow the user to toggle between the standard display mode set and the full display mode.

If you have set the standard display mode to hidden (which is also the default) or masked, the user can press Ctrl+R to change the display mode to Full, and Ctrl+R again to change it back to the standard one.

Obviously, if you have set the standard display mode to Full, pressing Ctrl+R wonโt cause any changes.

Confirmation: By default, the password will have a confirmation flow where the user will be asked for the input twice and the two responses will be compared. If they differ, an error message is shown and the user is prompted again.

By default, a โConfirmation:โ message is shown for the confirmation prompts, but this can be modified by setting a custom confirmation message only shown the second time, using the with_custom_confirmation_message() method.

If confirmation is not desired, it can be turned off using the without_confirmation() method.

Help message: Message displayed at the line below the prompt.

Formatter: Custom formatter in case you need to pre-process the user input before showing it as the final answer.

By default, it prints eight asterisk characters: ********.

Validators: Custom validators to make sure a given submitted input pass the specified requirements, e.g. not allowing empty inputs or requiring special characters.

No validators are on by default.

Remember that for CLI applications it is standard to not allow use any display modes other than Hidden and to not allow the user to see the text input in any way. Use the customization options at your discretion.

Example

 use inquire::{validator::{StringValidator, Validation}, Password, PasswordDisplayMode};

 let validator = |input: &str| if input.chars().count() < 10 {
     Ok(Validation::Invalid("Keys must have at least 10 characters.".into()))
 } else {
     Ok(Validation::Valid)
 };

 let name = Password::new("Encryption Key:")
     .with_display_toggle_enabled()
     .with_display_mode(PasswordDisplayMode::Hidden)
     .with_custom_confirmation_message("Encryption Key (confirm):")
     .with_custom_confirmation_error_message("The keys don't match.")
     .with_validator(validator)
     .with_formatter(&|_| String::from("Input received"))
     .with_help_message("It is recommended to generate a new one only for this purpose")
     .prompt();

 match name {
     Ok(_) => println!("This doesn't look like a key."),
     Err(_) => println!("An error happened when asking for your key, try again later."),
 }

Fields
message: &'a str

Message to be presented to the user.

custom_confirmation_message: Option<&'a str>

Message to be presented to the user when confirming the input.

custom_confirmation_error_message: Option<&'a str>

Error to be presented to the user when password confirmation fails.

help_message: Option<&'a str>

Help message to be presented to the user.

formatter: StringFormatter<'a>

Function that formats the user input and presents it to the user as the final rendering of the prompt.

display_mode: PasswordDisplayMode

How the password input is displayed to the user.

enable_display_toggle: bool

Whether to allow the user to toggle the display of the current password input by pressing the Ctrl+R hotkey.

enable_confirmation: bool

Whether to ask for input twice to see if the provided passwords are the same.

validators: Vec<Box<dyn StringValidator>>

Collection of validators to apply to the user input.

Validators are executed in the order they are stored, stopping at and displaying to the user only the first validation error that might appear.

The possible error is displayed to the user one line above the prompt.

render_config: RenderConfig<'a>

RenderConfig to apply to the rendered interface.

Note: The default render config considers if the NO_COLOR environment variable is set to decide whether to render the colored config or the empty one.

When overriding the config in a prompt, NO_COLOR is no longer considered and your config is treated as the only source of truth. If you want to customize colors and still support NO_COLOR, you will have to do this on your end.

Implementations

impl<'a> Password<'a>

pub const DEFAULT_FORMATTER: StringFormatter<'a>

Default formatter, set to always display "********" regardless of input length.

pub const DEFAULT_VALIDATORS: Vec<Box<dyn StringValidator>>

Default validators added to the Password prompt, none.

pub const DEFAULT_HELP_MESSAGE: Option<&'a str> = None

Default help message.

pub const DEFAULT_ENABLE_DISPLAY_TOGGLE: bool = false

Default value for the allow display toggle variable.

pub const DEFAULT_ENABLE_CONFIRMATION: bool = true

Default value for the enable confirmation variable.

pub const DEFAULT_DISPLAY_MODE: PasswordDisplayMode = PasswordDisplayMode::Hidden

Default password display mode.

pub fn new(message: &'a str) -> Self

Creates a Password with the provided message and default options.

pub fn with_help_message(self, message: &'a str) -> Self

Sets the help message of the prompt.

pub fn with_display_toggle_enabled(self) -> Self

Sets the flag to enable display toggling.

pub fn without_confirmation(self) -> Self

Disables the confirmation step of the prompt.

pub fn with_custom_confirmation_message(self, message: &'a str) -> Self

Sets the prompt message when asking for the password confirmation.

pub fn with_custom_confirmation_error_message(self, message: &'a str) -> Self

Sets the prompt error message when password confirmation fails.

pub fn with_display_mode(self, mode: PasswordDisplayMode) -> Self

Sets the standard display mode for the prompt.

pub fn with_formatter(self, formatter: StringFormatter<'a>) -> Self

Sets the formatter.

pub fn with_validator<V>(self, validator: V) -> Self
where V: StringValidator + 'static,

Adds a validator to the collection of validators. You might want to use this feature in case you need to limit the user to specific choices, such as requiring special characters in the password.

Validators are executed in the order they are stored, stopping at and displaying to the user only the first validation error that might appear.

The possible error is displayed to the user one line above the prompt.

pub fn with_validators(self, validators: &[Box<dyn StringValidator>]) -> Self

Adds the validators to the collection of validators in the order they are given.

Validators are executed in the order they are stored, stopping at and displaying to the user only the first validation error that might appear.

The possible error is displayed to the user one line above the prompt.

pub fn with_render_config(self, render_config: RenderConfig<'a>) -> Self

Sets the provided color theme to this prompt.

Note: The default render config considers if the NO_COLOR environment variable is set to decide whether to render the colored config or the empty one.

When overriding the config in a prompt, NO_COLOR is no longer considered and your config is treated as the only source of truth. If you want to customize colors and still support NO_COLOR, you will have to do this on your end.

pub fn prompt_skippable(self) -> InquireResult<Option<String>>

Parses the provided behavioral and rendering options and prompts the CLI user for input according to the defined rules.

This method is intended for flows where the user skipping/cancelling the prompt - by pressing ESC - is considered normal behavior. In this case, it does not return Err(InquireError::OperationCanceled), but Ok(None).

Meanwhile, if the user does submit an answer, the method wraps the return type with Some.

pub fn prompt(self) -> InquireResult<String>

Parses the provided behavioral and rendering options and prompts the CLI user for input according to the defined rules.