Trait inquire::InnerAction

pub trait InnerAction
where
    Self: Sized + Copy + Clone + PartialEq + Eq,
{
    type Config;

    // Required method
    fn from_key(key: Key, config: &Self::Config) -> Option<Self>
       where Self: Sized;
}

InnerActions are specialized prompt actions.

They must provide an implementation to optionally derive an action from a key event.

Required Associated Types

type Config

Configuration type for the prompt.

This is used to derive the action from a key event.

Required Methods

fn from_key(key: Key, config: &Self::Config) -> Option<Self>
where Self: Sized,

Derives a prompt action from a Key event and the prompt configuration.

Object Safety

This trait is not object safe.