inquire::ui

Struct RenderConfig

pub struct RenderConfig<'a> {

    pub prompt_prefix: Styled<&'a str>,
    pub answered_prompt_prefix: Styled<&'a str>,
    pub prompt: StyleSheet,
    pub default_value: StyleSheet,
    pub placeholder: StyleSheet,
    pub help_message: StyleSheet,
    pub password_mask: char,
    pub text_input: StyleSheet,
    pub answer: StyleSheet,
    pub canceled_prompt_indicator: Styled<&'a str>,
    pub error_message: ErrorMessageRenderConfig<'a>,
    pub highlighted_option_prefix: Styled<&'a str>,
    pub scroll_up_prefix: Styled<&'a str>,
    pub scroll_down_prefix: Styled<&'a str>,
    pub selected_checkbox: Styled<&'a str>,
    pub unselected_checkbox: Styled<&'a str>,
    pub option_index_prefix: IndexPrefix,
    pub option: StyleSheet,
    pub selected_option: Option<StyleSheet>,

}

Rendering configuration that can be applied to a prompt.

Render configurations can set mostly style sheets for particular parts of the prompt layout. Additionally, it allows you to set the content of a few tokens, such as prompt or error message prefixes.

Example

use inquire::ui::{Color, RenderConfig, Styled};

let empty: RenderConfig = RenderConfig::empty();
let default: RenderConfig = RenderConfig::default();

let prompt_prefix = Styled::new("$").with_fg(Color::DarkRed);
let mine = default.with_prompt_prefix(prompt_prefix);

Fields
prompt_prefix: Styled<&'a str>

Prefix added before prompts.

Note: a space character will be added to separate the prefix and the prompt message.

answered_prompt_prefix: Styled<&'a str>

Prefix added before answered prompts.

Note: a space character will be added to separate the prefix and the prompt message.

prompt: StyleSheet

Style of the prompt message, applicable to all prompt types.

default_value: StyleSheet

Render configuration of default values.

Note: default values are displayed wrapped in parenthesis, e.g. (yes). Non-styled space characters is added before the default value display and after the default value, as separators.

placeholder: StyleSheet

Render configuration of placeholders.

Note: placeholders are displayed wrapped in parenthesis, e.g. (yes). Non-styled space characters is added before the default value display and after the default value, as separators.

help_message: StyleSheet

Render configuration of help messages.

Note: help messages are displayed wrapped in brackets, e.g. [Be careful!].

password_mask: char

Character used to mask password text inputs when in mode Masked.

Note: Styles for masked text inputs are set in the text_input configuration.

text_input: StyleSheet

Style sheet for text inputs.

Note: a non-styled space character is added before the text input as a separator from the prompt message (or default value display).

answer: StyleSheet

Render configuration of final prompt answers (submissions).

Note: a non-styled space character is added before the answer as a separator from the prompt message (or default value display).

canceled_prompt_indicator: Styled<&'a str>

Render configuration of the message printed in the place of an answer when the prompt is canceled by the user - by pressing ESC.

Note: a non-styled space character is added before the indicator as a separator from the prompt message.

error_message: ErrorMessageRenderConfig<'a>

Render configuration for error messages.

highlighted_option_prefix: Styled<&'a str>

Prefix for the current highlighted option.

Note: a space character will be added to separate the prefix and the option value or the checkbox.

scroll_up_prefix: Styled<&'a str>

Prefix for the option listed at the top of the page, when it is possible to scroll up.

Note: a space character will be added to separate the prefix and the option value or the checkbox.

scroll_down_prefix: Styled<&'a str>

Prefix for the option listed at the bottom of the page, when it is possible to scroll down.

Note: a space character will be added to separate the prefix and the option value or the checkbox.

selected_checkbox: Styled<&'a str>

Selected checkbox in multi-select options.

Note: a space character will be added to separate the checkbox from a possible prefix, and to separate the checkbox from the option value to the right.

unselected_checkbox: Styled<&'a str>

Unselected checkbox in multi-select options.

Note: a space character will be added to separate the checkbox from a possible prefix, and to separate the checkbox from the option value to the right.

option_index_prefix: IndexPrefix

Definition of index prefixes in option lists.

option: StyleSheet

Style sheet for options.

Note: a non-styled space character is added before the option value as a separator from the prefix.

selected_option: Option<StyleSheet>

Style sheet for the option that is currently selected. If the value is None, it will fall back to option.

Note: a non-styled space character is added before the option value as a separator from the prefix.

Implementations

impl<'a> RenderConfig<'a>

pub fn empty() -> Self

RenderConfig in which no colors or attributes are applied.

pub fn default_colored() -> Self

RenderConfig where default colors and attributes are applied.

pub fn with_prompt_prefix(self, prompt_prefix: Styled<&'a str>) -> Self

Sets the prompt prefix and its style sheet.

pub fn with_answered_prompt_prefix( self, answered_prompt_prefix: Styled<&'a str>, ) -> Self

Sets the answered prompt prefix and its style sheet.

pub fn with_text_input(self, text_input: StyleSheet) -> Self

Sets style for text inputs.

pub fn with_default_value(self, default_value: StyleSheet) -> Self

Sets the style sheet for default values.

pub fn with_help_message(self, help_message: StyleSheet) -> Self

Sets the style sheet for help messages.

pub fn with_answer(self, answer: StyleSheet) -> Self

Sets the style sheet for answers.

pub fn with_error_message( self, error_message: ErrorMessageRenderConfig<'a>, ) -> Self

Sets the render configuration for error messages.

pub fn with_highlighted_option_prefix( self, highlighted_option_prefix: Styled<&'a str>, ) -> Self

Sets the styled component for prefixes in highlighted options.

pub fn with_scroll_up_prefix(self, scroll_up_prefix: Styled<&'a str>) -> Self

Sets the styled component for prefixes in scroll-up indicators.

pub fn with_scroll_down_prefix( self, scroll_down_prefix: Styled<&'a str>, ) -> Self

Sets the styled component for prefixes in scroll-down indicators.

pub fn with_selected_checkbox(self, selected_checkbox: Styled<&'a str>) -> Self

Sets the styled component for selected checkboxes.

pub fn with_unselected_checkbox( self, unselected_checkbox: Styled<&'a str>, ) -> Self

Sets the styled component for unselected checkboxes.

pub fn with_option_index_prefix(self, index_prefix: IndexPrefix) -> Self

Sets the index prefix for option lists.

pub fn with_option(self, option: StyleSheet) -> Self

Sets the style sheet for option values.

pub fn with_selected_option(self, selected_option: Option<StyleSheet>) -> Self

Sets the style sheet for currently selected option.

pub fn with_canceled_prompt_indicator( self, canceled_prompt_indicator: Styled<&'a str>, ) -> Self

Sets the indicator for canceled prompts.